refactor(api): move metrics initialization into api crate
Move metrics provider and endpoint setup into the top-level
carbide-api composition crate. Transfer JoinSet ownership with it so
process tasks are awaited outside api-core while core service
initialization remains unchanged.

Preserve the existing metric views and regression coverage, and remove
metrics-bootstrap dependencies that api-core no longer uses.
