+++ crates/api-core/Cargo.toml
@@ -41,7 +41,6 @@ carbide-dpf = { path = "../dpf" }
 carbide-firmware = { path = "../firmware" }
 carbide-health-metrics = { path = "../health-metrics" }
 carbide-health-report = { path = "../health-report" }
-carbide-http-connector = { path = "../http-connector" }
 carbide-instrument = { path = "../instrument" }
 carbide-host-support = { path = "../host-support", default-features = false }
 carbide-ib-fabric = { path = "../ib-fabric" }
@@ -77,7 +76,6 @@ component-manager = { path = "../component-manager" }
 config-version = { path = "../config-version", features = ["sqlx"] }
 dns-record = { path = "../dns-record" }
 libnmxc = { path = "../libnmxc" }
-metrics-endpoint = { path = "../metrics-endpoint" }
 mqttea = { path = "../mqttea" }
 nras = { path = "../nras" }
 spancounter = { path = "../spancounter" }
@@ -120,10 +118,6 @@ mac_address = { workspace = true }
 num_cpus = { workspace = true }
 nv-redfish = { workspace = true }
 opentelemetry = { workspace = true, features = ["logs"] }
-opentelemetry-prometheus = { workspace = true }
-opentelemetry-semantic-conventions = { workspace = true, features = [
-  "semconv_experimental",
-] }
 opentelemetry_sdk = { workspace = true, features = [
   "logs",
   "rt-tokio",
@@ -132,7 +126,6 @@ opentelemetry_sdk = { workspace = true, features = [
 ] }
 pkcs1 = { workspace = true }
 p256 = { workspace = true }
-prometheus = { workspace = true }
 rand = { workspace = true }
 regex = { workspace = true }
 reqwest = { workspace = true, default-features = false, features = [
