+++ crates/api/Cargo.toml
@@ -39,11 +39,15 @@ path = "src/main.rs"
 carbide-api-core = { path = "../api-core", default-features = false }
 carbide-api-db = { path = "../api-db", default-features = false }
 carbide-api-web = { path = "../api-web" }
+carbide-http-connector = { path = "../http-connector" }
 carbide-instrument = { path = "../instrument" }
 carbide-machine-controller = { path = "../machine-controller" }
+carbide-metrics-utils = { path = "../metrics-utils" }
 carbide-secrets = { path = "../secrets" }
+carbide-site-explorer = { path = "../site-explorer" }
 carbide-version = { path = "../version" }
 logfmt = { path = "../logfmt" }
+metrics-endpoint = { path = "../metrics-endpoint" }
 spancounter = { path = "../spancounter" }
 sqlx-query-tracing = { path = "../sqlx-query-tracing" }
 
@@ -53,12 +57,17 @@ eyre = { workspace = true }
 ipnetwork = { workspace = true }
 opentelemetry = { workspace = true, features = ["logs"] }
 opentelemetry-otlp = { workspace = true, features = ["grpc-tonic"] }
+opentelemetry-prometheus = { workspace = true }
+opentelemetry-semantic-conventions = { workspace = true, features = [
+  "semconv_experimental",
+] }
 opentelemetry_sdk = { workspace = true, features = [
   "logs",
   "rt-tokio",
   "spec_unstable_metrics_views",
   "testing",
 ] }
+prometheus = { workspace = true }
 sqlx = { workspace = true, features = [
   "runtime-tokio",
   "tls-rustls",
